requirements_train.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
langchain>=0.0.200
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.10.0
python-dotenv>=1.0.0
tqdm>=4.65.0